refactor tomopy code, add licence, permalink and specified line codes
