+++ imswitch/imcommon/model/mathutils.py
@@ -1,197 +1,6 @@
 import numpy as np
 
 
-#########################################################
-# 3D volumes operations, reused from the tomopy package #
-# https://tomopy.readthedocs.io/en/stable/              #
-#########################################################
-def _totuple(size, dim):
-    """
-    Converts size to tuple.
-    """
-    if not isinstance(size, tuple):
-        if dim == 2:
-            size = (size, size)
-        elif dim == 3:
-            size = (size, size, size)
-    return size
-
-
-def shepp3d(size=128, dtype='float32'):
-    """
-    Load 3D Shepp-Logan image array.
-
-    Parameters
-    ----------
-    size : int or tuple, optional
-        Size of the 3D data.
-    dtype : str, optional
-        The desired data-type for the array.
-
-    Returns
-    -------
-    ndarray
-        Output 3D test image.
-    """
-    size = _totuple(size, 3)
-    shepp_params = _array_to_params(_get_shepp_array())
-    return phantom(size, shepp_params, dtype).clip(0, np.inf)
-
-
-def phantom(size, params, dtype='float32'):
-    """
-    Generate a cube of given size using a list of ellipsoid parameters.
-
-    Parameters
-    ----------
-    size: tuple of int
-        Size of the output cube.
-    params: list of dict
-        List of dictionaries with the parameters defining the ellipsoids
-        to include in the cube.
-    dtype: str, optional
-        Data type of the output ndarray.
-
-    Returns
-    -------
-    ndarray
-        3D object filled with the specified ellipsoids.
-    """
-    # instantiate ndarray cube
-    obj = np.zeros(size, dtype=dtype)
-
-    # define coords
-    coords = _define_coords(size)
-
-    # recursively add ellipsoids to cube
-    for param in params:
-        _ellipsoid(param, out=obj, coords=coords)
-    return obj
-
-
-def _ellipsoid(params, shape=None, out=None, coords=None):
-    """
-    Generate a cube containing an ellipsoid defined by its parameters.
-    If out is given, fills the given cube instead of creating a new one.
-    """
-    # handle inputs
-    if shape is None and out is None:
-        raise ValueError("You need to set shape or out")
-    if out is None:
-        out = np.zeros(shape)
-    if shape is None:
-        shape = out.shape
-    if len(shape) == 1:
-        shape = shape, shape, shape
-    elif len(shape) == 2:
-        shape = shape[0], shape[1], 1
-    elif len(shape) > 3:
-        raise ValueError("input shape must be lower or equal to 3")
-    if coords is None:
-        coords = _define_coords(shape)
-
-    # rotate coords
-    coords = _transform(coords, params)
-
-    # recast as ndarray
-    coords = np.asarray(coords)
-    np.square(coords, out=coords)
-    ellip_mask = coords.sum(axis=0) <= 1.
-    ellip_mask.resize(shape)
-
-    # fill ellipsoid with value
-    out[ellip_mask] += params['A']
-    return out
-
-
-def _rotation_matrix(p):
-    """
-    Defines an Euler rotation matrix from angles phi, theta and psi.
-    """
-    cphi = np.cos(np.radians(p['phi']))
-    sphi = np.sin(np.radians(p['phi']))
-    ctheta = np.cos(np.radians(p['theta']))
-    stheta = np.sin(np.radians(p['theta']))
-    cpsi = np.cos(np.radians(p['psi']))
-    spsi = np.sin(np.radians(p['psi']))
-    alpha = [[cpsi * cphi - ctheta * sphi * spsi,
-              cpsi * sphi + ctheta * cphi * spsi,
-              spsi * stheta],
-             [-spsi * cphi - ctheta * sphi * cpsi,
-              -spsi * sphi + ctheta * cphi * cpsi,
-              cpsi * stheta],
-             [stheta * sphi,
-              -stheta * cphi,
-              ctheta]]
-    return np.asarray(alpha)
-
-
-def _define_coords(shape):
-    """
-    Generate a tuple of coords in 3D with a given shape.
-    """
-    mgrid = np.lib.index_tricks.nd_grid()
-    cshape = np.asarray(1j) * shape
-    x, y, z = mgrid[-1:1:cshape[0],
-                    -1:1:cshape[1],
-                    -1:1:cshape[2]]
-    return x, y, z
-
-
-def _transform(coords, p):
-    """
-    Apply rotation, translation and rescaling to a 3-tuple of coords.
-    """
-    alpha = _rotation_matrix(p)
-    out_coords = np.tensordot(alpha, coords, axes=1)
-    _shape = (3,) + (1,) * (out_coords.ndim - 1)
-    _dt = out_coords.dtype
-    M0 = np.array([p['x0'], p['y0'], p['z0']], dtype=_dt).reshape(_shape)
-    sc = np.array([p['a'], p['b'], p['c']], dtype=_dt).reshape(_shape)
-    out_coords -= M0
-    out_coords /= sc
-    return out_coords
-
-
-def _get_shepp_array():
-    """
-    Returns the parameters for generating modified Shepp-Logan phantom.
-    """
-    shepp_array = [
-        [1.,  .6900, .920, .810,   0.,     0.,   0.,   90.,   90.,   90.],
-        [-.8, .6624, .874, .780,   0., -.0184,   0.,   90.,   90.,   90.],
-        [-.2, .1100, .310, .220,  .22,     0.,   0., -108.,   90.,  100.],
-        [-.2, .1600, .410, .280, -.22,     0.,   0.,  108.,   90.,  100.],
-        [.1,  .2100, .250, .410,   0.,    .35, -.15,   90.,   90.,   90.],
-        [.1,  .0460, .046, .050,   0.,     .1,  .25,   90.,   90.,   90.],
-        [.1,  .0460, .046, .050,   0.,    -.1,  .25,   90.,   90.,   90.],
-        [.1,  .0460, .023, .050, -.08,  -.605,   0.,   90.,   90.,   90.],
-        [.1,  .0230, .023, .020,   0.,  -.606,   0.,   90.,   90.,   90.],
-        [.1,  .0230, .046, .020,  .06,  -.605,   0.,   90.,   90.,   90.]]
-    return shepp_array
-
-
-def _array_to_params(array):
-    """
-    Converts list to a dictionary.
-    """
-    # mandatory parameters to define an ellipsoid
-    params_tuple = [
-        'A',
-        'a', 'b', 'c',
-        'x0', 'y0', 'z0',
-        'phi', 'theta', 'psi']
-
-    array = np.asarray(array)
-    out = []
-    for i in range(array.shape[0]):
-        tmp = dict()
-        for k, j in zip(params_tuple, list(range(array.shape[1]))):
-            tmp[k] = array[i, j]
-        out.append(tmp)
-    return out
-
-
 #############################################################
 # Conversion of angles to steps and vice versa for Rotators #
 #############################################################
